requirements.txt
redis>=5.0.1
pydantic>=2.9.2
flask>=3.0.0
aiohttp>=3.10.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
black>=23.0.0
ruff>=0.1.0
safety>=3.0.0
pip-audit>=2.10.0
pyjwt>=2.8.0
bcrypt>=4.0.0
flask-limiter>=3.5.0
